move files; fix tests;
