+++ Quiz/Quiz.sln
@@ -25,13 +25,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Bot", "Bot", "{F97ECF76-96D
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Service", "Service", "{04B404F3-8DEA-4FFE-A691-718CD0E8EEAF}"
 EndProject
-
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuizLevelManager", "QuizLevelManager\QuizLevelManager.csproj", "{4EC01D8D-7BBC-4754-AB8F-9782DC3CC5AA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuizLevelManager", "QuizLevelManager\QuizLevelManager.csproj", "{4EC01D8D-7BBC-4754-AB8F-9782DC3CC5AA}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "LevelEditor", "LevelEditor", "{5DF206BE-B4B6-4691-A1BF-B3C3CB93C43E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuizRequestExtendedService", "QuizRequestExtendedService\QuizRequestExtendedService.csproj", "{01FB3E00-FAE2-41B0-9FB1-9D35DDE1989A}"
-
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuizRequestExtendedService", "QuizRequestExtendedService\QuizRequestExtendedService.csproj", "{01FB3E00-FAE2-41B0-9FB1-9D35DDE1989A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -151,7 +149,6 @@ Global
 		{9E18DF2D-93E4-4D66-BAA9-8BBF0A581B50}.Release|x64.Build.0 = Release|Any CPU
 		{9E18DF2D-93E4-4D66-BAA9-8BBF0A581B50}.Release|x86.ActiveCfg = Release|Any CPU
 		{9E18DF2D-93E4-4D66-BAA9-8BBF0A581B50}.Release|x86.Build.0 = Release|Any CPU
-
 		{4EC01D8D-7BBC-4754-AB8F-9782DC3CC5AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4EC01D8D-7BBC-4754-AB8F-9782DC3CC5AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4EC01D8D-7BBC-4754-AB8F-9782DC3CC5AA}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -176,7 +173,6 @@ Global
 		{01FB3E00-FAE2-41B0-9FB1-9D35DDE1989A}.Release|x64.Build.0 = Release|Any CPU
 		{01FB3E00-FAE2-41B0-9FB1-9D35DDE1989A}.Release|x86.ActiveCfg = Release|Any CPU
 		{01FB3E00-FAE2-41B0-9FB1-9D35DDE1989A}.Release|x86.Build.0 = Release|Any CPU
-
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -184,7 +180,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{CD831071-9E7D-42F8-B8A3-66D3114AAAC1} = {04B404F3-8DEA-4FFE-A691-718CD0E8EEAF}
 		{4B658940-58FA-4946-8192-A0F365048A7A} = {04B404F3-8DEA-4FFE-A691-718CD0E8EEAF}
-		{956E6D7B-3E2A-4FAE-B2A7-13BD0F66A93A} = {04B404F3-8DEA-4FFE-A691-718CD0E8EEAF}
 		{C6381A16-3CA1-4028-9D2F-181ECED7D252} = {04B404F3-8DEA-4FFE-A691-718CD0E8EEAF}
 		{DD59BF00-9CBF-4BB8-981F-51A394DC4989} = {F97ECF76-96D4-4D21-A7C7-2919DB160FB2}
 		{7D3A122D-AB28-4FAD-AC7C-1A9CD02FE975} = {F97ECF76-96D4-4D21-A7C7-2919DB160FB2}
